+++ api/src/evaluation/application/http-error-mapper-configuration.js
@@ -1,6 +1,6 @@
 import { HttpErrors } from '../../shared/application/http-errors.js';
 import { DomainErrorMappingConfiguration } from '../../shared/application/models/domain-error-mapping-configuration.js';
-import { CompetenceResetError, EmptyAnswerError, ImproveCompetenceEvaluationForbiddenError } from '../domain/errors.js';
+import { CompetenceResetError, EmptyAnswerError, ImproveCompetenceEvaluationForbiddenError, AcquiredBadgeForbiddenUpdateError } from '../domain/errors.js';
 
 const evaluationDomainErrorMappingConfiguration = [
   {
@@ -19,6 +19,12 @@ const evaluationDomainErrorMappingConfiguration = [
     name: CompetenceResetError.name,
     httpErrorFn: (error) => {
       return new HttpErrors.PreconditionFailedError(error.message);
+    },
+  },
+  {
+    name: AcquiredBadgeForbiddenUpdateError.name,
+    httpErrorFn: (error) => {
+      return new HttpErrors.ForbiddenError(error.message);
     },
   },
 ].map((domainErrorMappingConfiguration) => new DomainErrorMappingConfiguration(domainErrorMappingConfiguration));
